pyproject.toml
[project]
name = "content-enrichment-pipeline"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "structlog>=24.1.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "celery[redis]>=5.3.6",
  "redis>=5.0.8",
  "instructor>=1.3.2",
  "openai>=1.43.0",
  "crewai>=0.30.0",
  "pytest>=8.2.0",
]

[tool.uv]
package = true
